Add option to abbreviate labels for small window
Especially when the dock is located on the left or the right, user would
want to make the dock window as small as possible to gain the preview
and program view sizes. This option reduces the width of the labels so
that the dock consumes less area of the screen.